pyproject.toml
[project]
name = "customer360-demo"
version = "1.0.0"
description = "Customer 360 demo with ClickHouse Cloud and PuppyGraph"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Customer360 Team", email = "team@company.com"}
]
license = {text = "MIT"}
keywords = ["customer360", "graph-database", "clickhouse", "puppygraph", "analytics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    # ClickHouse integration
    "clickhouse-driver>=0.2.6",
    "clickhouse-connect>=0.7.0",
    
    # Data generation and processing
    "faker>=24.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "pyarrow>=15.0.0",
    "dask[complete]>=2024.1.0",
    
    # Graph database integration
    "neo4j>=5.17.0",  # For Cypher queries to PuppyGraph
    "py2neo>=2021.2.4",  # Alternative Neo4j client
    
    # Web application
    "streamlit>=1.31.0",
    "plotly>=5.19.0",
    "pyvis>=0.3.2",
    "streamlit-agraph>=0.0.45",  # Graph visualization component
    
    # Utilities
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.2",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "pydantic>=2.6.1",  # For data validation
    "typer>=0.9.0",  # CLI interface
    
    # Scientific computing
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    
    # Development and testing
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.61.0",
    "line-profiler>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/customer360-demo"
Repository = "https://github.com/yourorg/customer360-demo.git"
Documentation = "https://github.com/yourorg/customer360-demo/docs"
Issues = "https://github.com/yourorg/customer360-demo/issues"

[project.scripts]
customer360-generate = "data_generator.cli:app"
customer360-ingest = "ingestion.cli:app"
customer360-app = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]